Break config out into multiple files again because of puppet limitations.
Puppet does not seem to provide a simple (from vagrant's perspective) way to
make files attached to the main manifest accessible via the vagrant file
server. Additionally, gnome3 configuration sucks and was the cause of a ton
of frustration. Now the lockscreen doesn't appear after the VM idles for 300
seconds. Yay.